% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigViewer.R
\name{ConfigViewer}
\alias{ConfigViewer}
\title{ConfigViewer}
\usage{
ConfigViewer(pvalues, snp_probs = NULL, cond_pvalues = NULL,
  config_probs = NULL, correlations = NULL, z_file = NULL,
  ld_file = NULL, pval_threshold = 1e-06, snp_bf_threshold = 100,
  topconfigs = 5, subsample = TRUE, width = NULL, height = NULL)
}
\arguments{
\item{pvalues}{An obligatory data frame containing the column names 'rsid', 
'position', and 'pvalue'. Other columns may be included, but
will be ignored in analysis.}

\item{snp_probs}{An optional data frame containing the column names 'rsid',
'snp_prob', and 'snp_log10bf'. Corresponds to the .snp
output file from FINEMAP.}

\item{cond_pvalues}{An optional data frame containing the column names 'rsid',
and 'cond_pvalue'. Other columns may be included, but will be 
ignored in the analysis- meaning position values come from the pvalues
object.}

\item{config_probs}{An optional data frame containing the column names 
'rank', 'config', 'config_prob', and 'config_log10bf'. Corresponds to 
the .config output file from FINEMAP}

\item{correlations}{An optional data frame containing the column names
'rsid' and 'corr', giving each rsid's correlation to the top SNP. For 
datasets with a large LD correlation matrix, creating this object
    speeds up plot time significantly. Provide either correlations object,
    or z_file and ld_file together.}

\item{z_file}{An optional data frame containing the column name 'rsid'. 
Provide either correlations object, or z_file and ld_file together.
Corresponds to the .z file input to FINEMAP.}

\item{ld_file}{An optional correlations matrix of SNPs. Number of rows and
columns must match the z_file number of rows. Corresponds to the .ld
file input to FINEMAP.}

\item{pval_threshold}{Threshold for plot color coding and threshold line
placement of conventional manhattan plots. Default 1e-6.}

\item{snp_bf_threshold}{Threshold for plot color coding and threshold line 
placement of FINEMAP manhattan plot. Default 100, corresponding to a 
log10BF of 2.}

\item{topconfigs}{Number of top configurations to display.}

\item{subsample}{Logical value asking whether, for datasets containing more
than 5000 variants, the nonsignificant (in p value and bayes factor)
variants should be subsampled. Speeds up plot responsiveness significantly
    for large datasets.}

\item{width}{Manual setting of plot width- will disable dynamic resizing.}

\item{height}{Manual setting of plot height- will disable dynamic resizing.}
}
\description{
Interactive plots for simultaneously viewing conventional 
GWAS manhattan plots alongside FINEMAP results, with graphical
representation of most probable causal SNP configurations.
You can plot one to three manhattan plots simultaneously,
with only conventinal GWAS results (pvalues object) required.
For viewing manhattan plots with color-coded correlation to top  
SNP, you may upload either the z file and ld file used for 
FINEMAPing, or a correlations file with columns 'rsid' and 'corr'
representing rsids and correlations to top SNP, respectively.
For large ld matrices, creating this correlations data frame will
speed up plot generation time significantly.
}

